refactor(specmap): move specmap implementation to resolve module
Refs #3386
