+++ src/resolver/apidom/reference/dereference/strategies/openapi-3-1-swagger-client/errors/index.js
@@ -1,4 +1,4 @@
-import createError from '../../../../../../../specmap/lib/create-error.js';
+import createError from '../../../../../../specmap/lib/create-error.js';
 
 // eslint-disable-next-line import/prefer-default-export
 export const SchemaRefError = createError('SchemaRefError', function cb(message, extra, oriError) {
