+++ src/resolver/specmap/lib/all-of.js
