+++ src/resolver/specmap/lib/context-tree.js
