+++ src/resolver/specmap/lib/index.js
