+++ src/resolver/specmap/lib/properties.js
