+++ test/resolver/specmap/all-of.js
@@ -1,6 +1,6 @@
 import * as undici from 'undici';
 
-import mapSpec, { plugins } from '../../src/specmap/index.js';
+import mapSpec, { plugins } from '../../../src/resolver/specmap/index.js';
 
 describe('allOf', () => {
   let mockAgent;
