+++ test/resolver/specmap/data/complex/complex-example.json
