+++ test/resolver/specmap/data/cyclic/external/1.js
