+++ test/resolver/specmap/data/cyclic/external/10.js
