+++ test/resolver/specmap/data/cyclic/external/2.js
