+++ test/resolver/specmap/data/cyclic/external/20.js
