+++ test/resolver/specmap/data/cyclic/external/21.js
