+++ test/resolver/specmap/data/cyclic/external/30.js
