+++ test/resolver/specmap/data/cyclic/external/31.js
