+++ test/resolver/specmap/data/cyclic/external/32.js
