+++ test/resolver/specmap/data/cyclic/external/40.js
