+++ test/resolver/specmap/data/cyclic/internal/1.js
