+++ test/resolver/specmap/data/cyclic/internal/10.js
