+++ test/resolver/specmap/data/cyclic/internal/11.js
