+++ test/resolver/specmap/data/cyclic/internal/12.js
