+++ test/resolver/specmap/data/cyclic/internal/2.js
