+++ test/resolver/specmap/data/cyclic/internal/20.js
