+++ test/resolver/specmap/data/cyclic/internal/21.js
