+++ test/resolver/specmap/data/cyclic/internal/22.js
