+++ test/resolver/specmap/data/cyclic/internal/3.js
