+++ test/resolver/specmap/data/cyclic/internal/30.js
