+++ test/resolver/specmap/data/cyclic/internal/31.js
