+++ test/resolver/specmap/data/cyclic/internal/40.js
