+++ test/resolver/specmap/data/cyclic/internal/41.js
