+++ test/resolver/specmap/data/cyclic/internal/50.js
