+++ test/resolver/specmap/data/specs/example.json
