+++ test/resolver/specmap/data/specs/example.resolved.json
