+++ test/resolver/specmap/data/specs/petstore-simple-resolved.json
