+++ test/resolver/specmap/data/specs/petstore-simple.json
